move web/handlers/utils and web/handlers/service to web/
